+++ .eslintrc.js
@@ -9,6 +9,14 @@ module.exports = {
     'no-warning-comments': 'off',
     // A bunch of tools do wonky stuff with Node
     'node/global-require': 'off',
+    'no-restricted-syntax': [
+      'error',
+      {
+        selector:
+          'ImportDeclaration[specifiers.length=0][importKind=value][source.value!=/core-js.*/]',
+        message: "Use 'import type' for empty imports",
+      },
+    ],
   },
   overrides: [
     {