+++ packages/plugin-css/src/plugin-css.ts
@@ -11,8 +11,8 @@ import {
   TargetRuntime,
 } from '@sewing-kit/plugins';
 
-import {} from '@sewing-kit/plugin-jest';
-import {} from '@sewing-kit/plugin-webpack';
+import type {} from '@sewing-kit/plugin-jest';
+import type {} from '@sewing-kit/plugin-webpack';
 
 import {
   PostcssPlugins,