+++ packages/plugin-vscode/CHANGELOG.md
@@ -7,6 +7,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fix
+
+- Use explicit type imports for configuration hooks [[#144](https://github.com/Shopify/sewing-kit-next/pull/144/files)
+
 ## [0.1.14] - 2021-04-14
 
 - No updates. Transitive dependency bump.